+++ spec/helpers.rb
@@ -1,0 +1,3 @@
+def rendering_failure_message(title:, embed_code:)
+  "Unable to render block '#{title}' #{embed_code}"
+end